+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 - Bump `proguard` from v5.5.0 to v5.6.1
   This also bumps the proguard cache version from v3 to v4. ([#1734](https://github.com/getsentry/symbolicator/pull/1734))
 
+### Various fixes & improvements
+
+- Symbolicated JVM frames are now marked according to whether their method was
+  synthesized by the compiler. ([#1735](https://github.com/getsentry/symbolicator/pull/1735))
+
 ## 25.7.0
 
 ### Various fixes & improvements